New scripts to plot results
Modified significantly the scripts to get the results in folder "3".
Now, extracting the data and plotting it are clearly separated.
In addition, the data is now stored in easier-to-parse JSON files.

The plotting scripts are now in the `outputs` subfolder, do not assume
anymore that WIEN2K is the default code to compare with, and allow to
show results compared with one other plugin.

(In the future one can extend the scripts to compare with multiple plugins,
if one wants to - probably better done as a different script to avoid making
the logic too complex?)