Modifying scripts for formation energy
Now they can also compute the minimum-energy configuration of unaries
They still need some assessment of correctness

I checked and, with the current data (probably not very good?)
I get for
Pb-X|X2O|O
a large error of 6.878987 for the formation energy
(the two values being: -7.728713295810546 vs -0.8497267791535705)

Indeed, if I look at the values I see:
Pb-X2O (3 atoms/simulation)
  FLEUR: "E0": -1141042.0321028265
  WIEN : "E0": -1141021.3817287854

Lowest-energy Pb unary -> FCC
Pb-X/FCC (1 atom/simulation)
 FLEUR "E0": -569487.4147870957
 WIEN  "E0": -569487.4066672832

Lowest-energy O unary -> SC
O-X/SC (1 atom/simulation)
 WIEN  "E0": -2044.019213881717
 FLEUR "E0": -2044.016388747669

So the unaries agree quite well, but there are 21eV (!!)
of difference between the oxide calculation of FLEUR and WIEN2K.
Divided by 3 atoms, it is indeed ~7eV/atom, so the script seems correect.

Interestingly, if one computes the formation energies using *only*
oxides and endpoints, the results are not so bad.
So probably one of the two codes has a "consistent" shift for all oxides,
that is not applied for the unaries.

It would be interesting to see if this is still the case with the most recent data.