Some details required after the end of the verification paper.
The verification paper is now pubblished and the aiida-common-workflow package
has been released with version 1.0.1 making available all the code required
to reproduce the paper results.
We then update some parts here to comply with the last decisions of the projects
(e.g. use of the weights for the calculatio of nu) and the update due to the release
of aiida-common-workflows (e.g. the name of the protocol and the package version).