Removing the launch scripts for cp2k and gpaw
We decided to not keep the launching scripts for the various codes.
We believe that the quantum-espresso example (plus READEME explanation)
is sufficient for the users to understand the steps to run the calculation with any other
code. Moreover, most of the parameters are machine/profile dependent, not so many
are general for a given code.