First version of magnetic verification
Here we add the first subset of structures and instructions to run magnetic verifications by
performing fixed spin moment calculations to analyze the energy as a function of the (fixed) total magnetization.

In this commit, we provide the first set of structure `PBE-magnetic-v1`, which contains all configurations that result in a magnetic minimum for
the elements with 3d electrons and `O, F, Cl, Br`.