First version of magnetic verification: `PBE-magnetic-d-block-OFClBr-v1` (#52)
Here we add the first subset of structures and instructions to run magnetic verifications by
performing fixed spin moment calculations to analyze the energy as a function of the (fixed) total magnetization.

In this commit, we provide the first set of structure `PBE-magnetic-d-block-OFClBr-v1`, which contains all configurations that result in a magnetic minimum for
the elements from the d-block and `O, F, Cl, Br`.

We describe the verification procedure and define explicit inputs, i.e., which total magnetizations to run.